requirements.txt
annotated-types==0.7.0
attrs==23.2.0
Authlib==1.3.1
bandit==1.7.9
boltons==21.0.0
bracex==2.4
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-option-group==0.5.6
colorama==0.4.6
cryptography==43.0.0
defusedxml==0.7.1
Deprecated==1.2.14
dparse==0.6.4b0
exceptiongroup==1.2.2
face==22.0.0
filelock==3.12.4
glom==22.1.0
googleapis-common-protos==1.63.2
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
packaging==24.1
pbr==6.0.0
peewee==3.17.6
protobuf==4.25.4
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyGithub==2.3.0
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
safety==3.2.4
safety-schemas==0.0.2
semgrep==1.81.0
shellingham==1.5.4
stevedore==5.2.0
tomli==2.0.1
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
wcmatch==8.5.2
wrapt==1.16.0
zipp==3.19.2
